# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FrameX Studios API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Admin authentication
  - name: media
    description: Media management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: adminLogout
      tags: [auth]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getAuthMe
      tags: [auth]
      summary: Get current auth status
      responses:
        "200":
          description: Auth status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"

  /media:
    get:
      operationId: listMedia
      tags: [media]
      summary: List all media items
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [editing, thumbnails]
        - name: section
          in: query
          schema:
            type: string
            enum: [gaming, other]
      responses:
        "200":
          description: List of media
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaItem"

    post:
      operationId: createMedia
      tags: [media]
      summary: Upload a media item (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMediaRequest"
      responses:
        "201":
          description: Media created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /media/{id}:
    delete:
      operationId: deleteMedia
      tags: [media]
      summary: Delete a media item (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Media deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]

    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]

    AuthStatus:
      type: object
      properties:
        authenticated:
          type: boolean
      required: [authenticated]

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    MediaItem:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
          enum: [editing, thumbnails]
        section:
          type: string
          enum: [gaming, other]
        url:
          type: string
        type:
          type: string
          enum: [image, video, youtube]
        youtubeId:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, category, section, url, type, createdAt]

    CreateMediaRequest:
      type: object
      properties:
        category:
          type: string
          enum: [editing, thumbnails]
        section:
          type: string
          enum: [gaming, other]
        url:
          type: string
        type:
          type: string
          enum: [image, video, youtube]
        youtubeId:
          type: string
          nullable: true
      required: [category, section, url, type]